Separating map download progress to a component
Download bar is now per-map, and the progress bar is a re-usable component that just needs to be provided a map's "spring name". This component can be used elsewhere (e.g. in a lobby where the map is not downloaded yet).
Also fixed an error in the "parsedProgress" value coming up as NaN most of the time.