Map download progress added to Library > Map/Replay (#407)
This adds a small progress bar that appears when a map is being
downloaded from either the Library > Replays or Library > Maps > [id]
views. The progress bar appears below the buttons, and then disappears
when complete. It displays the progress as both a horizontal progress
bar as well as a two-digit number with percentage sign after (e.g.
"50%").

Contributed to https://github.com/beyond-all-reason/bar-lobby/issues/350