Replay screen updates (#422)
Fixes #406

The right panel of the Replay view now displays a tabbed interface. Two
tabs are presented when a replay is selected; Map and Details.

The **Map** tab will display the existing map preview with player
positions. It is the default view.

The **Details** tab will display:
* Engine version
*  Game version
* Team/player list 

We could extend this to include advanced replay details as well, if
desired in bar-lobby.

I have also corrected two issues with the "Download Map" button.
* The string "Download Map" was not styled to match the other versions
of the buttons. It now is properly styled.
* The string "Downloading Map..." was so large it would wrap to a second
line, causing the UI to jump around. This text only is now shrunk
slightly to fit on a single line to prevent this.