Rework i18n implementation to add typescript support (#429)
PR aims to add typescript support to the vue-i18n implementation.

#### Why?

While using the current implementation to start the translation of
hard-coded strings visible by the user, a few shortcomings where found
on the current implementation:

**Problem 1. No typescript support on tranlation function**
When replacing a string with a call to the translation function `{{
t(...) }}` the strings provided to `t(...)` need to be exactly the
correct translation key, this is workflow implies, that the developer
must search or add the translation in the source translation json files
and then build the translation key _(by hand)_ from the path of their
translation, by joining the keys with dots.

***For example***
If you want to add the translation for the _Quick play_ button first we
find the translation string:
```json
{
    "lobby": {
        "buttons": {
            "quickPlay": "Quick play"
        },
        ...
    }
}
```
Then replace the text `"Quick play"` with `{{
t("lobby.buttons.quickPlay") }}` in the `.vue` file.

As you can imagine that is not a maintainable workflow and very prone to
errors.

***Solution***
By adding typescript support, the typescript server can provide
autocompletions on translation keys directly.

In the example the string `"lobby.buttons.quickPlay"` would be suggested
by the autocompletions provided by ts-server.

**Problem 2. Git Submodule nested inside project**
It is planned in near future to have a git submodule to manage
translation files across projects. In the current implementation this
git submodule would live inside `src/renderer/assets/languages` this
introduces some amount of complexity when managing the future git
submodule.

***For example***
If a breaking change is added to the source translation files, it would
require a rollback of automated workflows to remove the breaking change.

***Solution***
Move source translation files to root `lang` directory and introduce an
additional build step for translation files for this repository.

Taking these problems into account these changes were made:

### Added

- New npm script `npm run generate-i18n-assets` that transforms
translation source files and generates their corresponding asset file in
`src/renderer/assets/languages`
- npm dev package [`tsx`](https://www.npmjs.com/package/tsx) for running
ts files as scripts
- New generated translation asset files in
`src/renderer/assets/languages`

### Changed

- Extracted the setup for `vue-i18n` to its own file
- Reworked the implementation of `vue-i18n` to load assets file directly
by importing them as modules and to add typescript support
- Moved source translation files to root `lang` directory

Contributes to #424 by adding to the i18n automated workflow
Contributes to #425 by adding the translation of
`src/renderer/views/home/overview.vue` as an example of usage