Overall improvement on integration with the recoil engine process (#436)
This is an overall improvement on integrating bar-lobby with spring :

1. bar-lobby was considering the game **launched** as soon as the spring
process is spawned. This is not good because spawning the process
doesn't guarantee the game will launch properly. I'm now waiting for
`"[Game::Load]"` on stdout before declaring the game **launched and
running**.

2. failure to launch the game wasn't properly propagating to the UI,
this PR fixes this issue as well.

3. Spring noticeably hangs for a few seconds upon exiting, even if game
isn't displayed anymore on screen. This PR adds a workaround by
considering the game **stopped** when we read `"[SpringApp::Kill][8]"`
on stdout.


Behavior change: starting the game hides the main window until the game
is exited.