Fetch and display dynamic matchmaking queues (#392) (#458)
Added fetchAvailableQueues which retrieves the matchmaking/list data.
This is called in window.tachyon.onConnected. Maybe there's a better
place for this but it seemed appropriate.

ranked.vue displays the dynamic playlist data now, and sorts it by total
player count

All of the queues are displayed using the 'classic' background image
defined in css currently as it doesn't look like there are specific
background images for multiplayer yet.

Closes #392