Fix tachyon-client request/response binding. (#465)
It was not using the generated tracking uuid for binding a response
handler to a specific request.

Causing issue when sending a bunch of the same commandId request. 
So for example, asking for 3 `user/info` in parallel, only two would get
their response handled, by a random responseHandler. The last response
would never get handled.