Replays Pagination Sorting Fix (#477)
Fixes: https://github.com/beyond-all-reason/bar-lobby/issues/468

Issue was identified with pagination on the replays page, where the
sorting was not working correctly.
This PR address the issue due to `.offset` and `.limit` being called
before sorting, which causes the sorting to only be applied to the 15
replays it was filtered to.
Updated the logic and cleaned up the code a bit here to get it working.