Updates to Download Button usage (#473)
With this, the Download button is fully shared across client in all
areas where it should be used. Fixes
https://github.com/beyond-all-reason/bar-lobby/issues/421

Fixed an newly-found issue where Quick Play drawer would not appear when
trying to launch it from Library>Map>[id]. There is still a problem,
however, where launching Quick Play from the Library>Map>[id] will error
out on loading. Recoil reports as an error that team 0 does not have a
leader 0. This appears to be a start-script creation issue and is being
investigated.