Component tweaks to support Online lobbies
Fixed a number of issues with sub-components that were designed with only offline battles in mind so they would play well with the current state of online lobbies.
Note that a number of features are completely disabled because there is no support for changing things in a lobby (except joining/leaving) in the Tachyon protocol yet.