Made ActiveLobby view dev-only
The "Active Lobby" view is now dev-mode visible only through a direct button to open it.
Default behavior now on joining is to open the BattleDrawer to the joined lobby.
Added a "Exit Lobby" button to the Lobby.
Added a Join/Open Lobby button that is reactive to the current selection.