Graceful disconnection
The entire client now handles disconnection better. User will be sent back to the login/play offline view if they disconnect while in an "online only" view.
Notifications instead of unhandled exceptions are used to advise the user of the errors.
Lobby will close and the custom lobby view will be in a "loading" state until the connection recovers or they are ejected to the login screen.
Pending server changes, the ``me.store`` will attempt to recover the currentLobby state upon reconnection.