Login Confirmation modal and Subnavigation (#505)
Fixes #497 and Fixes #498 

Created the Login confirmation modal, I used the button styles from
elsewhere in the app that matched he login button so it looked nice.
Used a composable so the state of the login modal could be accessed
easily from other components.

I decided to use "Login" and "Play Offline" to match what the login page
says and reuse it's translation text.
I made the onlineOnly UI items visible at all times, and then have them
open the modal when the user is not authenticated.
I added guards to the routing to open the modal when unauthenticated
users try to navigate to those views.