Fixed bot config desync
Bot settings now are correctly received and updated from the server, regardless of who submitted the changes.
If changes are received while the modal is open, user choices are "forgotten" and replaced with server-provided settings.
Also fixed it so if a user opens a modal, no request is sent to the server unless a change was made to the settings.