Matchmaking updated and cleaned up (#488)
This PR updates `matchmaking.store.ts`. The file has also been cleaned up
to be more readable, maintainable, and ready for future development.

Note that this does not implement any functionality that was not already
present, nor fix any functionality that was previously unimplemented.
This is just laying the groundwork for UI-side advancement of
Matchmaking, as well as matchmaking client logic.

As an aside, the main issue from matchmaking being functional is
ensuring the player has the map for their match downloaded before
needing to Ready.