DownloadContentButton can now handle Engine and Game download (#503)
Fixes #350 

## UX changes
``DownloadContentButton`` now has optional props of ``map``, ``engine``,
and ``game``. Behavior of the component is as follows:
1. For each of those properties, the component will determine if the
content is installed already. If any one (or more) property requires a
download, the button will become a "Download" button.
2. Clicking Download will result in an attempt to download only the
missing content. Downloads are consecutive, not concurrent, as
``pr-downloader`` always crashed for me if triggered multiple times.
3. The button will become a disabled button with an internal progress
bar overlay (as is already), moving from 0-100% for each item requested.
Once all requested content is downloaded, the button will revert to the
"no downloads required" state as currently in use in the client
(_usually_ implemented as "Play" button).
4. Failed downloads will result in errors in console and via
NotificationsAPI. The button will revert to the state as above in (1).
This will prevent users from gaining access to "Play" if they do not
have the required content and permit reattempts to download the content.

## Code changes/fixes
* ``downloadFailed`` was not being sent by ``pr-downloader``, this has
been fixed.
* Component ``MapDownloadProgress`` renamed to ``DownloadProgress``
* Component ``DownloadContentButton`` updated per UX changes above
* Download failure logic added to ``map.store`` ``engine.store`` and
``game.store``
* Failed downloads were not removed from ``downloads.store``, this has
been fixed.
* Updates to places where ``DownloadContentButton`` was already utilized
to utilize the UX changes above.
* ``test2.vue`` now has a testable map/engine/game
``DownloadContentButton``, which incidentally can be used to attempt a
forced download of specific map/engine/game content.